Портирование игры на другие операционные системы, проект собирается под MacOS WinXP32 Win64, заходит в игру показывает графику win64 и macos подключаются по DirectIP. В сетевой игре надписи Десинхронизация, Win64 может вылетать в сингл плеере,ассемблерные вставки переписаны на C для не Windows x86 платформ Много мусора и дебага в проекте WiP

+++ src/Main executable/Fastdraw.h
@@ -44,7 +44,7 @@ struct xRLCTable
 	int Size;
 	int sign;
 	int	SCount;
-	int OfsTable[32768000];
+	intptr_t OfsTable[32768000];
 };
 
 typedef xRLCTable* RLCTable;
@@ -74,7 +74,7 @@ public:
 
 	void SetStdShadow()
 	{
-		ShadowGP = int(RLC);
+		ShadowGP = (int)(intptr_t)(RLC);
 		ShadowDx = 1;
 		ShadowDy = 1;
 	};
